refactor: GameBoard를 ResultView로 클래스명 변경, 싱글톤으로 변경
